Fix difference for datetime on snapshot and replication in MySQL
In current version, we read the datetime from MySQL in the replication process and convert it to a time zone in the `ConnectionParams.Location`. At the same time we parse `datetime` in UTC location on snapshot. This fix starts to parse `datetime` on snapshot in `ConnectionParams.Locatio` too.

User problem:

The user's problem is that there is a datetime value of `2023-05-15 13:52:06` stored in MySQL. We read this event during the replication process and try to convert it into the time zone specified in the connection parameters. When we do so, we get `2023-05-15 13:52:06 +3`. Then, the CH destination converts the time to UTC, resulting in `2023-05-15 10:52:06 +0`.
commit_hash:c029a6e00302b8771242a8e6f50573f3088b135f